:recycle: REFACTOR: conversation/customer services into sunshine.ts
Also:

- auth_service renamed to authentication
