+++ lib/services/authentication.ts
