+++ specs/spec_helpers/fake_auth.ts
@@ -1,4 +1,4 @@
-import authService  from '../../lib/services/auth_service'
+import authService  from '../../lib/services/authentication'
 import { Staff }    from "@prisma/client";
 import sinon        from 'sinon'
